Move `snow.DefaultConsensusContextTest` to `snowtest.ConsensusContext` (#2507)
