+++ snow/consensus/snowman/consensus_test.go
@@ -20,9 +20,9 @@ import (
 	"gonum.org/v1/gonum/mathext/prng"
 
 	"github.com/ava-labs/avalanchego/ids"
-	"github.com/ava-labs/avalanchego/snow"
 	"github.com/ava-labs/avalanchego/snow/choices"
 	"github.com/ava-labs/avalanchego/snow/consensus/snowball"
+	"github.com/ava-labs/avalanchego/snow/snowtest"
 	"github.com/ava-labs/avalanchego/utils/bag"
 )
 
@@ -94,7 +94,7 @@ func InitializeTest(t *testing.T, factory Factory) {
 
 	sm := factory.New()
 
-	ctx := snow.DefaultConsensusContextTest()
+	ctx := snowtest.ConsensusContext()
 	params := snowball.Parameters{
 		K:                     1,
 		AlphaPreference:       1,
@@ -119,7 +119,7 @@ func NumProcessingTest(t *testing.T, factory Factory) {
 
 	sm := factory.New()
 
-	ctx := snow.DefaultConsensusContextTest()
+	ctx := snowtest.ConsensusContext()
 	params := snowball.Parameters{
 		K:                     1,
 		AlphaPreference:       1,
@@ -161,7 +161,7 @@ func AddToTailTest(t *testing.T, factory Factory) {
 
 	sm := factory.New()
 
-	ctx := snow.DefaultConsensusContextTest()
+	ctx := snowtest.ConsensusContext()
 	params := snowball.Parameters{
 		K:                     1,
 		AlphaPreference:       1,
@@ -200,7 +200,7 @@ func AddToNonTailTest(t *testing.T, factory Factory) {
 
 	sm := factory.New()
 
-	ctx := snow.DefaultConsensusContextTest()
+	ctx := snowtest.ConsensusContext()
 	params := snowball.Parameters{
 		K:                     1,
 		AlphaPreference:       1,
@@ -248,7 +248,7 @@ func AddToUnknownTest(t *testing.T, factory Factory) {
 
 	sm := factory.New()
 
-	ctx := snow.DefaultConsensusContextTest()
+	ctx := snowtest.ConsensusContext()
 	params := snowball.Parameters{
 		K:                     1,
 		AlphaPreference:       1,
@@ -288,7 +288,7 @@ func StatusOrProcessingPreviouslyAcceptedTest(t *testing.T, factory Factory) {
 
 	sm := factory.New()
 
-	ctx := snow.DefaultConsensusContextTest()
+	ctx := snowtest.ConsensusContext()
 	params := snowball.Parameters{
 		K:                     1,
 		AlphaPreference:       1,
@@ -317,7 +317,7 @@ func StatusOrProcessingPreviouslyRejectedTest(t *testing.T, factory Factory) {
 
 	sm := factory.New()
 
-	ctx := snow.DefaultConsensusContextTest()
+	ctx := snowtest.ConsensusContext()
 	params := snowball.Parameters{
 		K:                     1,
 		AlphaPreference:       1,
@@ -354,7 +354,7 @@ func StatusOrProcessingUnissuedTest(t *testing.T, factory Factory) {
 
 	sm := factory.New()
 
-	ctx := snow.DefaultConsensusContextTest()
+	ctx := snowtest.ConsensusContext()
 	params := snowball.Parameters{
 		K:                     1,
 		AlphaPreference:       1,
@@ -391,7 +391,7 @@ func StatusOrProcessingIssuedTest(t *testing.T, factory Factory) {
 
 	sm := factory.New()
 
-	ctx := snow.DefaultConsensusContextTest()
+	ctx := snowtest.ConsensusContext()
 	params := snowball.Parameters{
 		K:                     1,
 		AlphaPreference:       1,
@@ -430,7 +430,7 @@ func RecordPollAcceptSingleBlockTest(t *testing.T, factory Factory) {
 
 	sm := factory.New()
 
-	ctx := snow.DefaultConsensusContextTest()
+	ctx := snowtest.ConsensusContext()
 	params := snowball.Parameters{
 		K:                     1,
 		AlphaPreference:       1,
@@ -472,7 +472,7 @@ func RecordPollAcceptAndRejectTest(t *testing.T, factory Factory) {
 
 	sm := factory.New()
 
-	ctx := snow.DefaultConsensusContextTest()
+	ctx := snowtest.ConsensusContext()
 	params := snowball.Parameters{
 		K:                     1,
 		AlphaPreference:       1,
@@ -525,7 +525,7 @@ func RecordPollSplitVoteNoChangeTest(t *testing.T, factory Factory) {
 	require := require.New(t)
 	sm := factory.New()
 
-	ctx := snow.DefaultConsensusContextTest()
+	ctx := snowtest.ConsensusContext()
 	registerer := prometheus.NewRegistry()
 	ctx.Registerer = registerer
 
@@ -588,7 +588,7 @@ func RecordPollWhenFinalizedTest(t *testing.T, factory Factory) {
 
 	sm := factory.New()
 
-	ctx := snow.DefaultConsensusContextTest()
+	ctx := snowtest.ConsensusContext()
 	params := snowball.Parameters{
 		K:                     1,
 		AlphaPreference:       1,
@@ -613,7 +613,7 @@ func RecordPollRejectTransitivelyTest(t *testing.T, factory Factory) {
 
 	sm := factory.New()
 
-	ctx := snow.DefaultConsensusContextTest()
+	ctx := snowtest.ConsensusContext()
 	params := snowball.Parameters{
 		K:                     1,
 		AlphaPreference:       1,
@@ -683,7 +683,7 @@ func RecordPollTransitivelyResetConfidenceTest(t *testing.T, factory Factory) {
 
 	sm := factory.New()
 
-	ctx := snow.DefaultConsensusContextTest()
+	ctx := snowtest.ConsensusContext()
 	params := snowball.Parameters{
 		K:                     1,
 		AlphaPreference:       1,
@@ -775,7 +775,7 @@ func RecordPollInvalidVoteTest(t *testing.T, factory Factory) {
 
 	sm := factory.New()
 
-	ctx := snow.DefaultConsensusContextTest()
+	ctx := snowtest.ConsensusContext()
 	params := snowball.Parameters{
 		K:                     1,
 		AlphaPreference:       1,
@@ -816,7 +816,7 @@ func RecordPollTransitiveVotingTest(t *testing.T, factory Factory) {
 
 	sm := factory.New()
 
-	ctx := snow.DefaultConsensusContextTest()
+	ctx := snowtest.ConsensusContext()
 	params := snowball.Parameters{
 		K:                     3,
 		AlphaPreference:       3,
@@ -926,7 +926,7 @@ func RecordPollDivergedVotingTest(t *testing.T, factory Factory) {
 	sm := factory.New()
 	require := require.New(t)
 
-	ctx := snow.DefaultConsensusContextTest()
+	ctx := snowtest.ConsensusContext()
 	params := snowball.Parameters{
 		K:                     1,
 		AlphaPreference:       1,
@@ -1029,7 +1029,7 @@ func RecordPollDivergedVotingWithNoConflictingBitTest(t *testing.T, factory Fact
 	sm := factory.New()
 	require := require.New(t)
 
-	ctx := snow.DefaultConsensusContextTest()
+	ctx := snowtest.ConsensusContext()
 	params := snowball.Parameters{
 		K:                     1,
 		AlphaPreference:       1,
@@ -1133,7 +1133,7 @@ func RecordPollChangePreferredChainTest(t *testing.T, factory Factory) {
 
 	sm := factory.New()
 
-	ctx := snow.DefaultConsensusContextTest()
+	ctx := snowtest.ConsensusContext()
 	params := snowball.Parameters{
 		K:                     1,
 		AlphaPreference:       1,
@@ -1240,7 +1240,7 @@ func LastAcceptedTest(t *testing.T, factory Factory) {
 	sm := factory.New()
 	require := require.New(t)
 
-	ctx := snow.DefaultConsensusContextTest()
+	ctx := snowtest.ConsensusContext()
 	params := snowball.Parameters{
 		K:                     1,
 		AlphaPreference:       1,
@@ -1324,7 +1324,7 @@ func MetricsProcessingErrorTest(t *testing.T, factory Factory) {
 
 	sm := factory.New()
 
-	ctx := snow.DefaultConsensusContextTest()
+	ctx := snowtest.ConsensusContext()
 	params := snowball.Parameters{
 		K:                     1,
 		AlphaPreference:       1,
@@ -1353,7 +1353,7 @@ func MetricsAcceptedErrorTest(t *testing.T, factory Factory) {
 
 	sm := factory.New()
 
-	ctx := snow.DefaultConsensusContextTest()
+	ctx := snowtest.ConsensusContext()
 	params := snowball.Parameters{
 		K:                     1,
 		AlphaPreference:       1,
@@ -1382,7 +1382,7 @@ func MetricsRejectedErrorTest(t *testing.T, factory Factory) {
 
 	sm := factory.New()
 
-	ctx := snow.DefaultConsensusContextTest()
+	ctx := snowtest.ConsensusContext()
 	params := snowball.Parameters{
 		K:                     1,
 		AlphaPreference:       1,
@@ -1411,7 +1411,7 @@ func ErrorOnInitialRejectionTest(t *testing.T, factory Factory) {
 
 	sm := factory.New()
 
-	ctx := snow.DefaultConsensusContextTest()
+	ctx := snowtest.ConsensusContext()
 	params := snowball.Parameters{
 		K:                     1,
 		AlphaPreference:       1,
@@ -1450,7 +1450,7 @@ func ErrorOnAcceptTest(t *testing.T, factory Factory) {
 
 	sm := factory.New()
 
-	ctx := snow.DefaultConsensusContextTest()
+	ctx := snowtest.ConsensusContext()
 	params := snowball.Parameters{
 		K:                     1,
 		AlphaPreference:       1,
@@ -1487,7 +1487,7 @@ func ErrorOnRejectSiblingTest(t *testing.T, factory Factory) {
 
 	sm := factory.New()
 
-	ctx := snow.DefaultConsensusContextTest()
+	ctx := snowtest.ConsensusContext()
 	params := snowball.Parameters{
 		K:                     1,
 		AlphaPreference:       1,
@@ -1533,7 +1533,7 @@ func ErrorOnTransitiveRejectionTest(t *testing.T, factory Factory) {
 
 	sm := factory.New()
 
-	ctx := snow.DefaultConsensusContextTest()
+	ctx := snowtest.ConsensusContext()
 	params := snowball.Parameters{
 		K:                     1,
 		AlphaPreference:       1,
@@ -1623,7 +1623,7 @@ func ErrorOnAddDecidedBlockTest(t *testing.T, factory Factory) {
 	sm := factory.New()
 	require := require.New(t)
 
-	ctx := snow.DefaultConsensusContextTest()
+	ctx := snowtest.ConsensusContext()
 	params := snowball.Parameters{
 		K:                     1,
 		AlphaPreference:       1,
@@ -1653,7 +1653,7 @@ func ErrorOnAddDuplicateBlockIDTest(t *testing.T, factory Factory) {
 	sm := factory.New()
 	require := require.New(t)
 
-	ctx := snow.DefaultConsensusContextTest()
+	ctx := snowtest.ConsensusContext()
 	params := snowball.Parameters{
 		K:                     1,
 		AlphaPreference:       1,
@@ -1717,7 +1717,7 @@ func RecordPollWithDefaultParameters(t *testing.T, factory Factory) {
 
 	sm := factory.New()
 
-	ctx := snow.DefaultConsensusContextTest()
+	ctx := snowtest.ConsensusContext()
 	params := snowball.DefaultParameters
 	require.NoError(sm.Initialize(ctx, params, GenesisID, GenesisHeight, GenesisTimestamp))
