+++ snow/engine/snowman/config_test.go
@@ -4,18 +4,18 @@
 package snowman
 
 import (
-	"github.com/ava-labs/avalanchego/snow"
 	"github.com/ava-labs/avalanchego/snow/consensus/snowball"
 	"github.com/ava-labs/avalanchego/snow/consensus/snowman"
 	"github.com/ava-labs/avalanchego/snow/engine/common"
 	"github.com/ava-labs/avalanchego/snow/engine/common/tracker"
 	"github.com/ava-labs/avalanchego/snow/engine/snowman/block"
+	"github.com/ava-labs/avalanchego/snow/snowtest"
 	"github.com/ava-labs/avalanchego/snow/validators"
 )
 
 func DefaultConfig() Config {
 	return Config{
-		Ctx:                 snow.DefaultConsensusContextTest(),
+		Ctx:                 snowtest.ConsensusContext(),
 		VM:                  &block.TestVM{},
 		Sender:              &common.SenderTest{},
 		Validators:          validators.NewManager(),
