+++ snow/snowtest/snowtest.go
@@ -1,0 +1,30 @@
+// Copyright (C) 2019-2023, Ava Labs, Inc. All rights reserved.
+// See the file LICENSE for licensing terms.
+
+package snowtest
+
+import (
+	"github.com/prometheus/client_golang/prometheus"
+
+	"github.com/ava-labs/avalanchego/ids"
+	"github.com/ava-labs/avalanchego/snow"
+)
+
+var _ snow.Acceptor = noOpAcceptor{}
+
+type noOpAcceptor struct{}
+
+func (noOpAcceptor) Accept(*snow.ConsensusContext, ids.ID, []byte) error {
+	return nil
+}
+
+func ConsensusContext() *snow.ConsensusContext {
+	return &snow.ConsensusContext{
+		Context:             snow.DefaultContextTest(),
+		Registerer:          prometheus.NewRegistry(),
+		AvalancheRegisterer: prometheus.NewRegistry(),
+		BlockAcceptor:       noOpAcceptor{},
+		TxAcceptor:          noOpAcceptor{},
+		VertexAcceptor:      noOpAcceptor{},
+	}
+}
